+++ app/src/main/java/com/example/newagilityapp/data/repository/ProjectRepository.kt
@@ -8,12 +8,12 @@ import javax.inject.Inject
 
 class ProjectRepository @Inject constructor(private val projectDao: ProjectDao) {
     fun getProjectById(projectId: Int): Flow<Project?> = projectDao.getProjectById(projectId)
-    fun getTotalHoursFromSessions(projectId: Int): Flow<Int> = projectDao.getTotalHoursFromSessions(projectId)
+    fun getTotalHoursFromSessions(projectId: Int): Flow<Long> = projectDao.getTotalHoursFromSessions(projectId)
     fun getTotalProjectsCount(): Flow<Int> = projectDao.getTotalProjectCount()
     fun getCompletedProjectsCount(): Flow<Int> {
         return projectDao.getCompletedProjectsCount()
     }
-    fun getTotalHoursWorked(): Flow<Int> {
+    fun getTotalHoursWorked(): Flow<Long> {
         return projectDao.getTotalHoursWorked()
     }
     fun getTotalGoalHours(): Flow<Float> = projectDao.getTotalGoalHours()
